fix(build,ui,network,env_sensor): first-build pass + boot/reconnect hardening
Build fixes (first clean build, 1491/1491 steps pass):

- CMakeLists: remove device_hal/device (app_reset->button->led_strip 1.0 transitive conflict with new RMT RX used by env_sensor); remove examples/common; default IDF_TARGET to esp32c6

- sdkconfig.defaults: Matter Kconfig block (Test DAC Provider, Thread off, HKDF/IPv6 on, OTA on); disable DS at ESP-TLS + esp_secure_cert_mgr layers; replace ignored CONFIG_ENABLE_WIFI_AP with CONFIG_ESP_WIFI_SOFTAP_SUPPORT; drop unknown CONFIG_SPIRAM for C6

- matter_app.c: drop C++-only esp_matter.h includes (file is C)

- format specifiers: PRIu32 + (int) cast on uxTaskPriorityGet across 9 sites

- main/idf_component.yml: pin espressif/led_strip ^2.0.0 (override ^1.0.0); dependencies.lock committed for reproducible led_strip 2.5.5

Boot / reconnect hardening (Reviewer P1+P2):

- ui: led_strip 2.x API; ui_init does LED first then queue, full rollback; ui_is_initialized() lets main.c skip ui_task spawn (NO abort); long-press via 1-elem xQueueOverwrite (not volatile); rate-limited ui_write_rgb() checks set_pixel+refresh; ui.h self-contained

- main: UI failure non-fatal; sensors/state machine/network/Matter continue

- network: full Wi-Fi lifecycle - auto-connect on saved creds, async xTimer reconnect 1->60s backoff, auth-fail stops after 3 attempts, IP_GOT_IP resets state, network_init rolls back on failure

- env_sensor: DHT22 RMT RX late-callback race fixed - per-transaction ID, drain stale semaphore, rmt_disable() on timeout, discard late callbacks

Not yet validated: Matter remains stub; first build only proves compilation, NOT BLE commissioning / Matter reporting / OTA / sensor validation. Production blockers tracked in project_memory.

+++ firmware/components/ld2410c/ld2410c.c
@@ -59,7 +59,7 @@ static void sensor_radar_task(void *pv)
 
     ESP_LOGI(TAG, "task started (uart=%d, stack %u bytes, prio %d)",
              s_uart_num, (unsigned)uxTaskGetStackHighWaterMark(NULL),
-             uxTaskPriorityGet(NULL));
+             (int)uxTaskPriorityGet(NULL));
 
     for (;;) {
         int n = uart_read_bytes(s_uart_num, buf, sizeof(buf),
